OP_CLOSE (of OP_OPEN / normal capture) does not need to recurse. The backtracking can be handled entirely in OP_OPEN (the bounds will anyway not be valid until OP_OPEN restored the GrpStart, so we can defer GrpEnd untill then).

----

There were several occurrences of
```
     Result := MatchPrim(next);
     Exit;
```

If we continue the while-loop at the current level it will return the same result.



```
                                                           Time     Before |    OP_OPEN     |       exit     | Match count
==================================================================================================================
                                                  /Twain/ :          31 ms |      23 ms     |      23 ms     |       2388
                                              /(?i)Twain/ :          46 ms |      39 ms     |      47 ms     |       2657
                                             /[a-z]shing/ :         203 ms |     203 ms     |     203 ms     |       1877
                             /Huck[a-zA-Z]+|Saw[a-zA-Z]+/ :          23 ms |      31 ms     |      23 ms     |        396
                                                      /./ :         422 ms |     406 ms     |     398 ms     |   18905427
                                                    /(.)/ :         594 ms |     547 ms     |     547 ms     |   18905427
                                                      /e/ :          78 ms |      70 ms     |      78 ms     |    1781425
                                                    /(e)/ :          86 ms |      86 ms     |      86 ms     |    1781425
                                           /(?s).{1,45} / :          31 ms |      39 ms     |      31 ms     |     475715
                                         /(?s)\G.{1,45} / :          15 ms |      15 ms     |      16 ms     |      10616
                                         /\G(?s).{1,45} / :          16 ms |      15 ms     |      15 ms     |      10616
                                          /(?s).{1,45}? / :         148 ms |     156 ms     |     148 ms     |    3241534
                                        /(?s)\G.{1,45}? / :          15 ms |      15 ms     |      15 ms     |      69431
                                        /\G(?s).{1,45}? / :          16 ms |      15 ms     |      15 ms     |      69431
                                              /\b\w+nn\b/ :         203 ms |     195 ms     |     195 ms     |        359
                                       /[a-q][^u-z]{13}x/ :         609 ms |     609 ms     |     617 ms     |       4929
                            /Tom|Sawyer|Huckleberry|Finn/ :          31 ms |      31 ms     |      31 ms     |       3015
                        /(?i)Tom|Sawyer|Huckleberry|Finn/ :         125 ms |     133 ms     |     133 ms     |       4820
                    /.{0,2}(Tom|Sawyer|Huckleberry|Finn)/ :        1632 ms |    1656 ms     |    1679 ms     |       3015
                    /.{2,4}(Tom|Sawyer|Huckleberry|Finn)/ :        1859 ms |    1867 ms     |    1891 ms     |       2220
                      /Tom.{10,25}river|river.{10,25}Tom/ :          39 ms |      39 ms     |      46 ms     |          2
                                           /[a-zA-Z]+ing/ :         547 ms |     523 ms     |     539 ms     |      95863
                                  /\s[a-zA-Z]{0,12}ing\s/ :         242 ms |     242 ms     |     242 ms     |      67810
                          /([A-Za-z]awyer|[A-Za-z]inn)\s/ :         500 ms |     516 ms     |     515 ms     |        313
                              /["'][^"']{0,30}[?!\.]["']/ :          62 ms |      62 ms     |      62 ms     |       9857
                                /Tom(.{3,3}|.{5,5})*Finn/ :        2750 ms |    2703 ms     |    2695 ms     |         11
                                    /Tom(...|.....)*Finn/ :        1523 ms |    1132 ms +++ |    1125 ms     |         11
                                   /Tom(...|.....)*?Finn/ :        1336 ms |    1242 ms +++ |    1266 ms     |         11
                      /Tom((...|.....){2,9}\s){1,5}?Finn/ :        3593 ms |    3296 ms +++ |    3296 ms     |         11
                      /Tom((...|.....){2,9}?\s){1,5}Finn/ :        3586 ms |    3320 ms +++ |    3289 ms     |         11
                    /Tom((?:...|.....){2,9}\s){1,5}?Finn/ :        2851 ms |    2765 ms     |    2711 ms     |         11
                    /Tom((?:...|.....){2,9}?\s){1,5}Finn/ :        2812 ms |    2726 ms     |    2711 ms     |         11
                    /Tom((?>...|.....){2,9}\s){1,5}?Finn/ :          23 ms |      31 ms     |      31 ms     |          2
                    /Tom((?>...|.....){2,9}?\s){1,5}Finn/ :          31 ms |      23 ms     |      31 ms     |          2
                                        /\G(?is).(?=.*$)/ :         820 ms |     836 ms     |     804 ms +   |   20045118
                                /\G(?is).(?=(.){1,5}?$)?/ :        3101 ms |    2859 ms +++ |    2789 ms     |   20045118
                                       /\G(?is).(?=.*+$)/ :         836 ms |     875 ms     |     781 ms +++ |   20045118
                /\G(?is).{10,10}(?=(e|y|on|fin|.){0,20})/ :        1414 ms |    1234 ms +++ |    1250 ms     |    2004511
              /\G(?is).{10,10}(?=(?>e|y|on|fin|.){0,20})/ :        1406 ms |    1375 ms     |    1383 ms     |    2004511
                                          /Tom(?!.*Finn)/ :          23 ms |      31 ms     |      31 ms     |       2441
     /(?i)(?>[aeoui][bcdfghjklmnpqrstvwxyz \r\n]*){1,40}/ :         593 ms |     601 ms     |     594 ms     |     579843
                          /(?i)[ bdlm][abdegij][mnoprst]/ :         211 ms |     211 ms     |     210 ms     |     788955
Total:                                                            34491 ms |   32805 ms     |   32601 ms     |
```