@User4martin 2023-11-22
Added "implement \K"
https://regex101.com/r/tFDGt8/1

Any part before the \K will not be reported as part of the match. (It can contain captures / and it must be present)

For look around regex101 does either report it is not supported, or (seems? to) always fail the expression.
However, it makes perfect sense in look behind `bar|(?<=\K.)foo`  would start the match one before foo (yes could be written without the look behind in this case).
For look ahead, it can happen that the start is set to after the match ends `(?=...\K)a` => since a negative match len makes no sense, I decided that counts as a "not matched". 
Since regex101 reports errors (or non match) for this, the result does not conflict with any other engine.

Zero length matches are ok / also in regex101 https://regex101.com/r/O15wNt/1