@Alexey-T 2023-11-22
you are programming genius! Respect. I will commit soon.