@Alexey-T 2023-11-22
Can you add the docs text about \K, please?