@User4martin 2023-11-22
Before I commit, please review:

```
Matched Result
--------------

The begin of the match can be set using ``\K``.

By default the entire text covered by a pattern is considered matched. However it is possible to set explicitly what will be reported.

The pattern ``a\Kb`` will require the text to contain "ab". But only the "b" will be reported as having been matched.
Their can be several ``\K`` in a pattern, The last one will set the match-start position.
Only ``\K`` in active parts of the pattern are considered. E.g. ``a(\Kb)?`` will not consider ``\K`` if there is no "b".
Captures can exist outside the match set by ``\K``.
```